Small update
QR Code is now clickable, which redirects the user to the generated URL.

Minor visual changes and additions.

Still tinkering with changing the filename of the downloadable responses. It currently always saves as "responses.csv," which can clutter up the downloads folder and doesn't give a good idea of how recent the responses are.